+++ RealEstateProjectSaleBusinessObject/ViewModels/PaymentProcessDetailVM.cs
@@ -12,7 +12,7 @@ namespace RealEstateProjectSaleBusinessObject.ViewModels
         public string DetailName { get; set; }
         public string PeriodType { get; set; }
         public string? Period { get; set; }
-        public int? PaymentRate { get; set; }
+        public double? PaymentRate { get; set; }
         public string PaymentType { get; set; }
         public double Amount { get; set; }
         public string? Note { get; set; }
